requirements.txt
# Professional Certificate Generator - Python Dependencies
# Author: devag7 (Deva Garwalla)
# 
# Core Dependencies
celery==5.3.6
qrcode[pil]==7.4.2
python-dotenv==1.0.0
reportlab==4.1.0
Pillow==11.1.0
amqp==5.3.1
redis==5.0.1

# Additional Utilities
redis==5.0.1
kombu==5.4.2
click==8.1.8
typing-extensions==4.12.2

# Optional but recommended
psutil==5.9.6